[Engine]: make a distinction between static and physical objects
