+++ src/GameObjects/PhysicsGameObject.js
@@ -3,19 +3,11 @@ import { SceneManagerInstance } from "../Scenes/SceneManager.js";
 import { PositionPrevisionsInstance } from "../Utils/PositionPrevisions.js";
 import { CanvasInstance } from "../Utils/Canvas.js";
 
-export default class GameObject {
-	constructor(x, y, fillColor, strokeColor) {
-		this.id = Math.random();
-		this.name = `Obj - ${this.id}`;
+import GameObject from "./GameObjectBase.js";
 
-		// Render
-		this.x = x;
-		this.y = y;
-		this.z = 0;
-		this.lastPosition = { x: this.x, y: this.y, z: this.z };
-		this.fillColor = fillColor;
-		this.strokeColor = strokeColor;
-		this.visible = true;
+export default class PhysicsGameObject extends GameObject {
+	constructor(x, y, fillColor, strokeColor) {
+		super(x, y, fillColor, strokeColor);
 
 		// Physics
 		this.active = true;
@@ -28,53 +20,9 @@ export default class GameObject {
 		this._strokeDebugColor = "#016301";
 	}
 
-	setName(name) { this.name = name; }
-
-	// Render
-	// Position
-	setX(x) { this.setPosition(x, this.y, this.z); }
-	setY(y) { this.setPosition(this.x, y, this.z); }
-	setZ(z) { this.setPosition(this.x, this.y, z); }
-	getPosition() { return { x: this.x, y: this.y, z: this.z }; }
-	setPosition(x, y, z = this.z, force = false) {
-		if (this.bodyType === "S" && !force) return;
-		this.lastPosition = { x: this.x, y: this.y, z: this.z };
-
-		this.x = x;
-		this.y = y;
-		this.z = z;
-	}
-	setRandomPosition(x = 0, y = 0, width = GlobalStateManagerInstance.viewportDimensions.width, height = GlobalStateManagerInstance.viewportDimensions.height) {
-		do {
-			this.setPosition(
-				x + Math.random() * width,
-				y + Math.random() * height
-			);
-		} while (!this.checkIsInsideWorldBounds());
-	}
-
+	// -- Physics
 	getType() { return this._type; }
 
-	getCenter() { return { x: this.getCenterX(), y: this.getCenterY() }; }
-
-	getTopLeft() { return { x: this.getLeft(), y: this.getTop() }; }
-	getTopCenter() { return { x: this.getCenterX(), y: this.getTop() }; }
-	getTopRight() { return { x: this.getRight(), y: this.getTop() }; }
-
-	getBottomLeft() { return { x: this.getLeft(), y: this.getBottom() }; }
-	getBottomCenter() { return { x: this.getCenterX(), y: this.getBottom() }; }
-	getBottomRight() { return { x: this.getRight(), y: this.getBottom() }; }
-
-	getLeftCenter() { return { x: this.getLeft(), y: this.getCenterY() }; }
-	getRightCenter() { return { x: this.getRight(), y: this.getCenterY() }; }
-
-	// Color
-	setFillColor(fillColor) { this.fillColor = fillColor; }
-	setStrokeColor(strokeColor) { this.strokeColor = strokeColor; }
-
-	setVisible(isVisible) { this.visible = isVisible; }
-
-	// -- Physics
 	setActive(isActive) { this.active = isActive; }
 
 	// Body Type
@@ -166,14 +114,6 @@ export default class GameObject {
 			this.x + this.velocity.x * SceneManagerInstance.deltaTime,
 			this.y + this.velocity.y * SceneManagerInstance.deltaTime
 		);
-	}
-
-	_render() {
-		if (!this.visible) return;
-
-		CanvasInstance.context.fillStyle = this.fillColor;
-		CanvasInstance.context.strokeStyle = this.strokeColor;
-		this._renderType();
 	}
 
 	_debug() {
